fix: distribute scoring orchestration complexity
